refactor(flags): move cobra/config.go to flags
