+++ pkg/cmd/flags/config.go
@@ -1,4 +1,4 @@
-package cobra
+package flags 
 
 import (
 	"fmt"
