+++ pkg/cmd/flags/config_test.go
@@ -1,4 +1,4 @@
-package cobra
+package flags
 
 import (
 	"log"
